SAM/CODE/requirements.txt
absl-py==2.3.1
acvl_utils==0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
anytree==2.13.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
asyncio-atexit==1.0.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
batchgenerators==0.25
batchgeneratorsv2==0.2
beautifulsoup4==4.12.3
bigtree==1.0.0
bleach==6.1.0
Brotli @ file:///croot/brotli-split_1736182456865/work
bypy==1.8.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
comm==0.2.2
connected-components-3d==3.18.0
contourpy==1.2.1
cripser==0.0.21
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dicom2nifti==2.4.11
dill==0.3.9
docker-pycreds==0.4.0
docopt-ng==0.9.0
docstring_parser==0.17.0
dynamic_network_architectures==0.3.1
einops==0.8.0
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.20.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
googleapis-common-protos==1.63.2
graphviz==0.20.3
grpcio==1.75.1
gudhi==3.11.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imagecodecs==2024.6.1
imageio==2.34.2
imgaug==0.4.0
importlib-metadata==6.11.0
iopath==0.1.10
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.32.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kdtree==0.16
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
llvmlite==0.45.1
magicgui==0.10.1
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
MedPy==0.5.2
mistune==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
monai==1.3.2
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multiprocess==0.70.17
munch==4.0.0
narwhals==2.7.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nnunetv2==2.5.1
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.3.1
pooch==1.8.2
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.50
protobuf==4.25.4
psutil==5.9.8
psygnal==0.14.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyconify==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydicom==2.4.4
Pygments==2.19.1
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pysodmetrics==1.4.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
python-json-logger==2.0.7
pytz==2024.1
pyvista==0.46.3
PyYAML==6.0.1
pyzmq==26.2.1
QtPy==2.4.3
referencing==0.35.1
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
rtree==1.4.1
sacred==0.8.7
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
scooby==0.10.2
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.0.7
SimpleITK==2.3.1
six==1.17.0
skan==0.13.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
superqt==0.7.6
sympy==1.13.1
synapseclient==4.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
toolz==1.0.0
torch==2.4.0
torchaudio==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tornado==6.4.2
tqdm==4.66.4
traceback2==1.4.0
traitlets==5.14.3
triton==3.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2024.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==1.26.19
vtk==9.5.2
wandb==0.19.7
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.11
wrapt==1.16.0
yacs==0.1.8
zipp==3.21.0
